refactor(fs-bridge): replace error message in `resolveSafePath` with `BridgePathTraversal`
* Updated the error handling in `resolveSafePath` to throw a `BridgePathTraversal` error instead of a generic error message.
* This change enhances clarity and consistency in error reporting related to path traversal issues.
* Removed the `resolveSafePath` tests from `node.test.ts` and added comprehensive tests in a new `utils.test.ts` file.
